Fix for issue #2 [Link to the Issue](https://github.com/nparashar150/threejs_/issues/2#issue-1086621778)
Glitch in the Chats area can be fixed by forcefully unmouning and remounting the specific component (UserChat.js) by changing the key value in the component

I have added Loading Spinner when chaning the chats which appears when the component is fetching the specified Users' Chat from the database.

I have used Lazy Loading for showing "Loading..." on slower networks.